+++ console/registry_catalog_test.go
@@ -3,8 +3,8 @@ package console
 import (
 	"testing"
 
-	"github.com/zatrano/framework/v2/manifest"
-	"github.com/zatrano/framework/v2/registry"
+	"github.com/zatrano/framework/v2/distribution/manifest"
+	"github.com/zatrano/framework/v2/distribution/registry"
 )
 
 func TestEcosystemCatalogBuildsRegistryIndex(t *testing.T) {