+++ distribution/registry/resolve_test.go
@@ -4,7 +4,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/zatrano/framework/v2/manifest"
+	"github.com/zatrano/framework/v2/distribution/manifest"
 )
 
 func constrainedSession() Index {